adding map

+++ app/src/main/java/com/example/myapplication/data/model/Parking.kt
@@ -14,8 +14,8 @@ data class Parking(
     val pricePerHour: Double,
     val description: String,
     val rating: Float,
-    val longitude:Float,
-    val latitude:Float
+    val longitude:Double,
+    val latitude:Double
 )
